% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribs.R
\name{imageAttributes}
\alias{imageAttributes}
\alias{bidsToDivest}
\alias{bidsToDivest.list}
\alias{bidsToDivest.character}
\alias{bidsToDivest.default}
\alias{divestToBids}
\alias{divestToBids.list}
\title{Extended image attributes}
\usage{
imageAttributes(x)

bidsToDivest(x)

\method{bidsToDivest}{list}(x)

\method{bidsToDivest}{character}(x)

\method{bidsToDivest}{default}(x)

divestToBids(x)

\method{divestToBids}{list}(x)

\method{bidsToDivest}{default}(x)
}
\arguments{
\item{x}{An R object. For \code{imageAttributes} this would usually be an
image object, like those returned by \code{\link{readDicom}}.
\code{bidsToDivest} is S3 generic, with methods for lists (of named
attributes) and character strings (giving the path to a BIDS JSON file),
as well as a default method that handles image objects.
\code{divestToBids} is also S3 generic, but does not offer a method for
file names, since file storage is not standardised in this case.}
}
\value{
A list of image attributes, possibly with its naming convention
  changed relative to the input.
}
\description{
These functions extract and manipulate medical image attributes that go
beyond the core metadata associated with the NIfTI-1 file format.
}
\details{
The DICOM format can encapsulate copious amounts of metadata about the scan
and the patient, which can be useful for more advanced or research-focussed
post-processing methods. Some of this information is extracted during the
DICOM-to-NIfTI conversion process and stored in additional named attributes;
the \code{imageAttributes} function returns a list of just these extended
attributes. The other two functions convert between \code{divest}'s own
naming convention and that used by the BIDS standard.
}
\note{
Attributes may include sensitive or identifiable information such as
  a patient's name, sex, date of birth, etc., if this was included in the
  original DICOM files. These functions make no attempt to anonymise this
  metadata, and so this must be handled by the user if necessary.
}
\examples{
path <- system.file("extdata", "raw", package="divest")
images <- readDicom(path, interactive=FALSE)
imageAttributes(images[[1]])
}
\references{
More information about metadata captured in within the BIDS
  format can be found at \url{https://bids.neuroimaging.io} or in the paper
  cited below.

K.J. Gorgolewski, T. Auer, V.D. Calhoun, et al. The brain imaging data
structure, a format for organizing and describing outputs of neuroimaging
experiments (2016). Scientific Data 3:160044. \doi{10.1038/sdata.2016.44}.
}
\author{
Jon Clayden <code@clayden.org>
}
